LayoutTests/plugins/netscape-plugin-setwindow-size-2-expected.txt

CONSOLE MESSAGE: PLUGIN: NPP_SetWindow: 200 200
This tests that a hidden plug-in gets a correct NPP_SetWindow the first time.
